+++ CHANGELOG.md
@@ -26,11 +26,13 @@ All notable changes to this project will be documented in this file. This change
 	- `:zabb` - zero adjusted beta binomial
 	- `:zibb` - zero inflated beta binomial
 * `:half-normal` - half-normal distribution
-* `:truncated` distribution
+* `:truncated` and `:mixture` distributions
 * `hpdi-extent` and `pi-extent` - new extents based on HPDI and PI from `rethinking` R package
 * `distribution-parameters` accept a keyword additionally
 * `fastmath.solver/find-root` - root solver for univariate functions
 * `delta-eq` and `delta=` for checking eqality with given accuracy.
+* `coefficient-matrix` and `correlation-matrix`
+* grid functions accept separate coordinates
 
 ### Changed